Implementing new DB binary
This is a beta version which is not yet committed upstream to the main development branch of the defaultbrowser github repository.

It has been forked and improved slightly to provide some command line parameters to benefit third party utility use.

These changes have been submitted as a pull request to the defaultbrowser github repository.

Signed-off-by: Stuart Ryan <stuart@ryan.email>